delete bullshit folder
